refactor: main dukan ui state
